Increment version number to 1.2.2 Renamed "Play Online" to "Network Play" to better reflect its intended use (although playing online is theoretically possible) Completely revamped Network Play menu to be controller friendly as much as possible. Join IP, Join and Host Port, Max players and song selection rules are now saved to settings and remembered for subsequent network games New graphics for Network Play menu, including header and menu backgrounds

+++ Assets/Scripts/Common/SettingsManager.cs
@@ -49,11 +49,11 @@ public class SettingsManager : MonoBehaviour
     [Header("Network Settings")]
     public int NetGameHostMaxPlayers = 8;
     public NetSongSelectRules NetGameHostSongSelectRules = NetSongSelectRules.AnyonePicks;
-    public int NetGameHostPort = 3334;
+    public ushort NetGameHostPort = 3334;
     public string NetGameJoinIpAddress = "127.0.0.1";
-    public int NetGameJoinPort = 3334;
+    public ushort NetGameJoinPort = 3334;
 
-    public const int DEFAULT_PORT = 3334;
+    public const ushort DEFAULT_PORT = 3334;
 
     // Start is called before the first frame update
     void Awake()
